Remove the Zmanim object from the HDate object
Based on some discussion I read this would not be healthy as it creates a G-D object.
A smarter move would be instead to create either a third class that would wrap both, or else
even better might be to create a property that would instantiate a zmanim object and check
the given time in relationship to the times from the Zmanim object.

Another option would be to create it as a property of the Zmanim object which would instantiate
a HDate object.