Change and add erev chagim to all be of holiday_type == 2
Erev shavuot had a holiday type of 9 which doesn't match other holiday_type 9.
Other chagim didn't have any erev chagim specified except for erev yom kippur
which was holiday_type 2.

Unfortuantely with the current code, this doesn't simplify the check for issur_melacha
as in the case of diaspora the first day yom tov is holiday_type 1, maybe holiday type
should be a list instead of an int. Call it holiday properties. This would allow
hoshana raba to be defined as chol hamoed, erev yom tov and special.